Adding specs

+++ app/controllers/api/payments_controller.rb
@@ -1,11 +1,7 @@
 class Api::PaymentsController < Api::ApiController
     before_action :is_renter?, except: [:producthistory]
-    def new
-        payment=PaymentHistory.new
-        product=Product.find_by(id:params[:product])
-        rental=RentalHistory.find_by(id:params[:rental])
-    end
-
+    before_action :is_owner?, only:[:producthistory]
+    
     def create
         payment=PaymentHistory.new
         payment.cardnumber=params[:cardnumber]
@@ -13,14 +9,19 @@ class Api::PaymentsController < Api::ApiController
         payment.cvc=params[:cvc]
         payment.amount=params[:amount]
         rental=RentalHistory.find_by(id:params[:rental_id])
-        product=Product.find_by(id:rental.product_id)
-        product.applicants.delete_all   
-        if payment.save
-            render json: payment ,status: :ok
-        else
-            render json: { message: "Error while saving"}, status: :unprocessable_entity
+        if rental
+            product=Product.find_by(id:rental.product_id)
+            product.applicants.delete_all   
+            if payment.save
+                render json: payment ,status: :ok
+            else
+                render json: { message: "Error while saving"}, status: :unprocessable_entity
 
+            end
+        else
+            render json: { message: "No rentals Found"}, status: :not_found
         end
+
     end
 
     def show
@@ -45,16 +46,26 @@ class Api::PaymentsController < Api::ApiController
 
     def producthistory
         product=Product.find_by(id:params[:id])
-        rentals=product.rental_histories
-        if rentals
-            render json: rentals, status: :ok
+        if product
+            rentals=product.rental_histories
+            if rentals
+                render json: rentals, status: :ok
+            else
+                render json: {message:"Rental not found"}, status: :not_found
+            end
         else
-            render json: {message:"Rental not found"}, status: :not_found
+            render json: { message: "No Product Found"}, status: :not_found
+        end
     end
 
     private
     def is_renter?
         unless current_account&& current_account.renter?
+            render json: {message: "You are not authorized to view this page"} , status: :unauthorized
+        end
+    end
+    def is_owner?
+        unless current_account && current_account.user?            
             render json: {message: "You are not authorized to view this page"} , status: :unauthorized
         end
     end
